requirements.txt
aiogram==3.13.1
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
psycopg[binary]==3.2.1
PyYAML==6.0.2
google-generativeai==0.7.2
